+++ apps/checklistsF/models.py
@@ -66,7 +66,7 @@ class Transformer(models.Model):
     substation = models.ForeignKey(
         Substation, on_delete=models.CASCADE, related_name="transformers"
     )
-    power = models.PositiveIntegerField(_("Transformer Power"))
+    power = models.FloatField(_("Transformer Power"))
     impedance = models.FloatField(_("Impedance"), blank=True, null=True)
     demand = models.FloatField(_("Demand"), blank=True, null=True)
     type = models.CharField(_("Type"), choices=TYPE_CHOICES, max_length=3, blank=True, null=True)
